+++ src/component/main/GiftRankingList.tsx
@@ -8,11 +8,12 @@ import { useNavigate } from 'react-router-dom';
 import type { ProductItem } from '@/type/GiftAPI/product';
 import type { RankType, TargetType } from '@/type/giftRanking';
 import { CentorAlignDiv240, Gap } from '@/styles/CommomStyle/Common.styled';
-import { BrandImage, Price, ProductCard, ProductGrid, ProductImage, ProductInfo } from '@/styles/CommomStyle/ProductList';
+import {  ProductGrid,} from '@/styles/CommomStyle/ProductList';
 import Loading from '../Loading';
 import { baseRankingUrl } from '@/constant/api';
 import { getFromUrl } from '@/utils/getFromUrl';
 import { useQuery } from '@tanstack/react-query';
+import ProductCard from '../ProductCard';
 
 
 
@@ -32,20 +33,9 @@ const GiftRankingList = ({ targetType, rankType }: GiftRankingListProps) => {
         queryKey: ['rankingData', RankingUrl],
         queryFn: () => getFromUrl(RankingUrl)
     });
-    const { user } = useAuth();
-    const navigate = useNavigate();
 
     const visibleCount = isExpanded ? data?.length : GIFTLENGTH;
     const shownProducts = data ? (data as ProductItem[]).slice(0, visibleCount) : [];
-
-    const handleClickProduct = (item: ProductItem) => {
-        if (!user) {
-            navigate(`/login?redirect=/order?id=${item.id}`);
-        } else {
-            navigate(`/order?id=${item.id}`);
-        }
-    };
-
 
     if (error) return null
 
@@ -61,19 +51,8 @@ const GiftRankingList = ({ targetType, rankType }: GiftRankingListProps) => {
     return (
         <>
             <ProductGrid>
-                {shownProducts.map((item) => (
-                    <ProductCard
-                        key={item.id}
-                        onClick={() => handleClickProduct(item)}
-                    >
-                        <ProductImage src={item.imageURL} alt={item.name} />
-                        <BrandImage
-                            src={item.brandInfo.imageURL}
-                            alt={item.brandInfo.name}
-                        />
-                        <ProductInfo title={item.name}>{item.name}</ProductInfo>
-                        <Price>{item.price.sellingPrice.toLocaleString()} 원</Price>
-                    </ProductCard>
+                {shownProducts.map((product) => (
+                    <ProductCard {...product} />
                 ))}
             </ProductGrid>
             <Gap height={16} />
