+++ src/page/Product.tsx
@@ -1,0 +1,11 @@
+import { DefaultDiv } from "@/styles/Common.styled"
+
+const Product = () => {
+    return (
+        <DefaultDiv>
+
+        </DefaultDiv>
+    )
+}
+
+export default Product
